[ADD] Icon for agent on hacker map [MOD] Brighter wood floor

+++ Assets/ImageEffects/HeartRateMonitor/HeartRateShader.shader
@@ -42,7 +42,7 @@
 			sampler2D _MainTex;
 			sampler2D _HeartRateTex;
 			sampler2D _GridTex;
-			fixed4 _Color;
+			float4 _Color;
 			float4 _UVMult;
 
 			fixed4 frag (v2f i) : SV_Target
